Restore boxen-specific config for v2.2
This commit adjusts the LaunchCtl startup script to tell ElasticSearch to use the elasticsearch.yml file that’s tailored for Boxen. This config runs ElasticSearch on a non-standard port, in keeping with other Boxen services, and it sets the logging and data directories in a way that is more generally consistent with Boxen services. Note that this commit will change the ElasticSearch port. By default it runs on 9200. With this commit, it will run on $BOXEN_ELASTICSERCH_PORT.