[!] Upgrade ElasticSearch to v5.1
With this commit, we introduce a number of changes. First, we remove the
custom brew file and instead rely on Homebrew’s ElasticSearch package. We
update the default version to 5.1 instead of 2.2.1. We also remove the puppet-java
dependency, and instead install java using Homebrew Cask, now supported in
Boxen, which gives us Java 1.8, which we need for recent versions of
ElasticSearch. We still use the custom plist and ES config to keep it all boxeny.

Note that ES 2.x indexes are not compatible wtih ES 5.x indexes. You will have to
delete and recreate your indexes.